requirements.txt
Pillow==9.3.0
streamlit==1.16.0
torch==1.13.1
transformers==4.25.1
